refactor(runtime): move binding-context to observation
