+++ packages/runtime/src/binding/connectable.ts
@@ -1,6 +1,6 @@
 import { Class, IIndexable, Tracer } from '@aurelia/kernel';
 import { IBindingTargetObserver, IPropertySubscriber, LifecycleFlags } from '../observation';
-import { IObserverLocator } from '../observation//observer-locator';
+import { IObserverLocator } from '../observation/observer-locator';
 import { IBinding } from './binding';
 
 // TODO: add connect-queue (or something similar) back in when everything else is working, to improve startup time
