+++ packages/runtime/src/binding/interpolation-binding.ts
@@ -1,7 +1,7 @@
 import { IServiceLocator } from '@aurelia/kernel';
 import { IBindScope, State } from '../lifecycle';
 import { IBindingTargetAccessor, IScope, LifecycleFlags } from '../observation';
-import { IObserverLocator } from '../observation//observer-locator';
+import { IObserverLocator } from '../observation/observer-locator';
 import { IExpression, Interpolation } from './ast';
 import { IBinding, IBindingTarget } from './binding';
 import { BindingMode } from './binding-mode';
